fix(security): restore direct sync key rotation
Move the concrete sync store into storage so direct migration APIs can resolve and await it without startup registration while preserving the legacy import for one release.
