+++ aragora/security/migration.py
@@ -22,30 +22,15 @@ from collections.abc import Callable
 
 logger = logging.getLogger(__name__)
 
-SyncStoreProvider = Callable[[], Any]
 MigrationAuditProvider = Callable[..., Any]
 
-_sync_store_provider: SyncStoreProvider | None = None
 _migration_audit_provider: MigrationAuditProvider | None = None
-
-
-def register_sync_store_provider(provider: SyncStoreProvider | None) -> None:
-    """Register the higher-layer sync store provider."""
-    global _sync_store_provider
-    _sync_store_provider = provider
 
 
 def register_migration_audit_provider(provider: MigrationAuditProvider | None) -> None:
     """Register the higher-layer security audit provider."""
     global _migration_audit_provider
     _migration_audit_provider = provider
-
-
-def _get_registered_sync_store() -> Any:
-    """Return the configured sync store or preserve the former import failure."""
-    if _sync_store_provider is None:
-        raise ImportError("Sync store provider not registered")
-    return _sync_store_provider()
 
 
 def _audit_security(**kwargs: Any) -> Any:
@@ -354,40 +339,51 @@ def migrate_gmail_token_store(dry_run: bool = False) -> MigrationResult:
 
 def migrate_sync_store(dry_run: bool = False) -> MigrationResult:
     """Migrate connector sync store secrets."""
+    result = MigrationResult(store_name="sync_store")
+    sensitive_fields = [
+        "api_key",
+        "api_secret",
+        "token",
+        "password",
+        "auth_token",
+        "secret",
+    ]
+
     try:
-        store = _get_registered_sync_store()
-        migrator = EncryptionMigrator(dry_run=dry_run)
+        from aragora.storage.sync_store import get_sync_store
+        from aragora.utils.async_utils import run_async
 
-        # Connector credentials
-        sensitive_fields = [
-            "api_key",
-            "api_secret",
-            "token",
-            "password",
-            "auth_token",
-            "secret",
-        ]
+        store = run_async(get_sync_store())
+        connectors = run_async(store.list_connectors())
+        result.total_records = len(connectors)
 
-        def list_all():
-            if hasattr(store, "list_all"):
-                return list(store.list_all())
-            return []
+        for connector in connectors:
+            try:
+                if not needs_migration(connector.config, sensitive_fields):
+                    result.already_encrypted += 1
+                    continue
 
-        def save(job_id, record):
-            if hasattr(store, "save"):
-                return store.save(record)
-            return False
-
-        return migrator.migrate_store(
-            store_name="sync_store",
-            list_fn=list_all,
-            save_fn=save,
-            sensitive_fields=sensitive_fields,
-            id_field="job_id",
-        )
-    except ImportError as e:
+                if not dry_run:
+                    run_async(
+                        store.save_connector(
+                            connector.id,
+                            connector.connector_type,
+                            connector.name,
+                            connector.config,
+                        )
+                    )
+                result.migrated_records += 1
+            except (KeyError, ValueError, TypeError, RuntimeError, OSError) as e:
+                result.failed_records += 1
+                result.errors.append(f"Error migrating record: {connector.id}")
+                logger.warning("Failed to migrate sync connector %s: %s", connector.id, e)
+    except (ImportError, RuntimeError, ValueError, TypeError, OSError) as e:
         logger.warning("Sync store not available: %s", e)
-        return MigrationResult(store_name="sync_store", errors=["Sync store not available"])
+        result.errors.append("Sync store not available")
+
+    result.completed_at = datetime.now(timezone.utc)
+    result.duration_seconds = (result.completed_at - result.started_at).total_seconds()
+    return result
 
 
 @dataclass
@@ -538,6 +534,7 @@ def rotate_and_reencrypt_store(
     sensitive_fields: list[str],
     id_field: str = "id",
     encryption_service: Any | None = None,
+    force_resave: bool = False,
 ) -> MigrationResult:
     """
     Re-encrypt all records in a store with the current active key.
@@ -552,6 +549,7 @@ def rotate_and_reencrypt_store(
         sensitive_fields: Fields that are encrypted
         id_field: Name of the ID field in records
         encryption_service: Encryption service (uses global if not provided)
+        force_resave: Re-save plaintext records so the store encrypts them with the active key
 
     Returns:
         MigrationResult with statistics
@@ -577,6 +575,16 @@ def rotate_and_reencrypt_store(
             try:
                 # Check if record has encrypted fields
                 has_encrypted = any(is_field_encrypted(record.get(f)) for f in sensitive_fields)
+
+                if force_resave:
+                    if has_encrypted:
+                        raise RuntimeError("Store returned an encrypted record")
+                    if save_fn(record_id, record):
+                        result.migrated_records += 1
+                    else:
+                        result.failed_records += 1
+                        result.errors.append(f"Failed to save record: {record_id}")
+                    continue
 
                 if not has_encrypted:
                     # Record has no encrypted fields, skip
@@ -660,8 +668,9 @@ def rotate_encryption_key(
     service = get_encryption_service()
 
     # Get current key info
-    old_key = service._keys.get(service._active_key_id)
-    old_key_id = service._active_key_id or "default"
+    active_key_id = service._active_key_id
+    old_key = service._keys.get(active_key_id) if active_key_id else None
+    old_key_id = active_key_id or "default"
     old_version = old_key.version if old_key else 0
 
     result = KeyRotationResult(
@@ -720,15 +729,12 @@ def rotate_encryption_key(
             if not config_fn:
                 logger.warning("Unknown store for key rotation: %s", store_name)
                 continue
-            if store_name == "sync" and _sync_store_provider is None:
-                result.failed_records += 1
-                result.errors.append("Store sync re-encryption failed")
-                logger.error("Key rotation failed for sync: sync store provider not registered")
-                continue
-
             try:
                 config = config_fn()
                 if config is None:
+                    if store_name == "sync":
+                        result.failed_records += 1
+                        result.errors.append("Store sync re-encryption failed")
                     continue
 
                 store_result = rotate_and_reencrypt_store(
@@ -738,6 +744,7 @@ def rotate_encryption_key(
                     sensitive_fields=config["sensitive_fields"],
                     id_field=config["id_field"],
                     encryption_service=service,
+                    force_resave=config.get("force_resave", False),
                 )
 
                 result.stores_processed += 1
@@ -816,11 +823,37 @@ def _get_gmail_store_config() -> dict[str, Any] | None:
 def _get_sync_store_config() -> dict[str, Any] | None:
     """Get configuration for sync store re-encryption."""
     try:
-        store = _get_registered_sync_store()
+        from aragora.storage.sync_store import get_sync_store
+        from aragora.utils.async_utils import run_async
+
+        store = run_async(get_sync_store())
+        connectors_by_id: dict[str, Any] = {}
+
+        def list_connectors() -> list[dict[str, Any]]:
+            connectors = run_async(store.list_connectors())
+            connectors_by_id.clear()
+            records: list[dict[str, Any]] = []
+            for connector in connectors:
+                connectors_by_id[connector.id] = connector
+                records.append({**connector.config, "__sync_connector_id__": connector.id})
+            return records
+
+        def save_connector(connector_id: str, record: dict[str, Any]) -> bool:
+            connector = connectors_by_id[connector_id]
+            config = {key: value for key, value in record.items() if key != "__sync_connector_id__"}
+            run_async(
+                store.save_connector(
+                    connector.id,
+                    connector.connector_type,
+                    connector.name,
+                    config,
+                )
+            )
+            return True
 
         return {
-            "list_fn": lambda: list(store.list_all()) if hasattr(store, "list_all") else [],
-            "save_fn": lambda id, record: store.save(record) if hasattr(store, "save") else False,
+            "list_fn": list_connectors,
+            "save_fn": save_connector,
             "sensitive_fields": [
                 "api_key",
                 "api_secret",
@@ -829,11 +862,10 @@ def _get_sync_store_config() -> dict[str, Any] | None:
                 "auth_token",
                 "secret",
             ],
-            "id_field": "job_id",
+            "id_field": "__sync_connector_id__",
+            "force_resave": True,
         }
     except ImportError:
-        if _sync_store_provider is not None:
-            raise
         return None
 
 
@@ -842,7 +874,6 @@ __all__ = [
     "MigrationResult",
     "KeyRotationResult",
     "StartupMigrationConfig",
-    "register_sync_store_provider",
     "register_migration_audit_provider",
     "is_field_encrypted",
     "needs_migration",
